Access modal for calendars

+++ www/common/common-ui-elements.js
@@ -2921,7 +2921,7 @@ define([
 
         var dismiss = function () {
             common.mailbox.dismiss(data, function (err) {
-                console.log(err);
+                if (err) { console.log(err); }
             });
         };
         var answer = function (yes) {
@@ -2930,6 +2930,7 @@ define([
                 href: msg.content.href,
                 password: msg.content.password,
                 title: msg.content.title,
+                calendar: msg.content.calendar,
                 answer: yes
             }, {
                 channel: msg.content.user.notifications,
@@ -2965,16 +2966,29 @@ define([
                     // Add the pad to your drive
                     // This command will also add your mailbox to the metadata log
                     // The callback is called when the pad is stored, independantly of the metadata command
-                    sframeChan.query('Q_ACCEPT_OWNERSHIP', data, function (err, res) {
-                        if (err || (res && res.error)) {
-                            return void console.error(err | res.error);
-                        }
-                        UI.log(Messages.saved);
-                        if (autoStoreModal[data.channel]) {
-                            autoStoreModal[data.channel].delete();
-                            delete autoStoreModal[data.channel];
-                        }
-                    });
+                    if (data.calendar) {
+                        var calendarModule = common.makeUniversal('calendar');
+                        var calendarData = data.calendar;
+                        calendarData.href = data.href;
+                        calendarData.teamId = 1;
+                        calendarModule.execCommand('ADD', calendarData, function (obj) {
+                            if (obj && obj.error) {
+                                console.error(obj.error);
+                                return void UI.warn(Messages.error);
+                            }
+                        });
+                    } else {
+                        sframeChan.query('Q_ACCEPT_OWNERSHIP', data, function (err, res) {
+                            if (err || (res && res.error)) {
+                                return void console.error(err | res.error);
+                            }
+                            UI.log(Messages.saved);
+                            if (autoStoreModal[data.channel]) {
+                                autoStoreModal[data.channel].delete();
+                                delete autoStoreModal[data.channel];
+                            }
+                        });
+                    }
 
                     // Remove yourself from the pending owners
                     sframeChan.query('Q_SET_PAD_METADATA', {
